Refactor: modify directory and URL naming
